\name{limoaddin-package}
\alias{limoaddin-package}
\alias{limoaddin}
\docType{package}
\title{
\packageTitle{limoaddin}
}
\description{
\packageDescription{limoaddin}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{limoaddin}
\packageIndices{limoaddin}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{limoaddin}

Maintainer: \packageMaintainer{limoaddin}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
